[9.0] remove JobDB's site mask

+++ tests/Integration/ResourceStatusSystem/Test_SiteStatus.py
@@ -24,7 +24,6 @@ testSite = "test1234.test.test"
 @pytest.fixture(name="stClient")
 def fixtureSiteStatus():
     siteStatus = SiteStatus()
-    siteStatus.rssFlag = True
     yield siteStatus
 
 
@@ -58,25 +57,26 @@ def test_addAndRemove_simpleCase(stClient):
 
     result = stClient.getSites()
     assert result["OK"] is True, result["Message"]
-    assert testSite in result["Value"]
+    inRSS = testSite in result["Value"]
 
-    # TEST getSiteStatuses
-    # ...............................................................................
+    if inRSS:
+        # TEST getSiteStatuses
+        # ...............................................................................
 
-    result = stClient.getSiteStatuses([testSite])
-    assert result["OK"] is True, result["Message"]
-    assert result["Value"][testSite] == "Active"
+        result = stClient.getSiteStatuses([testSite])
+        assert result["OK"] is True, result["Message"]
+        assert result["Value"][testSite] == "Active"
 
-    # TEST getUsableSites
-    # ...............................................................................
+        # TEST getUsableSites
+        # ...............................................................................
 
-    result = stClient.getUsableSites([testSite])
-    assert result["OK"] is True, result["Message"]
-    assert testSite in result["Value"]
+        result = stClient.getUsableSites([testSite])
+        assert result["OK"] is True, result["Message"]
+        assert testSite in result["Value"]
 
-    # finally delete the test site
-    result = rsClient.deleteStatusElement("Site", "Status", testSite)
-    assert result["OK"] is True, result["Message"]
+        # finally delete the test site
+        result = rsClient.deleteStatusElement("Site", "Status", testSite)
+        assert result["OK"] is True, result["Message"]
 
 
 def test_addAndRemove_complicatedTest(stClient):
@@ -132,10 +132,7 @@ def test_addAndRemove_complicatedTest(stClient):
 
     result = stClient.getSites()
     assert result["OK"] is True, result["Message"]
-
-    assert "testActive1.test.test" in result["Value"]
-    assert "testActive.test.test" in result["Value"]
-    assert "testBanned.test.test" not in result["Value"]
+    inRSS = "testBanned.test.test" in result["Value"]
 
     # TEST getSites
     # ...............................................................................
@@ -143,25 +140,17 @@ def test_addAndRemove_complicatedTest(stClient):
     result = stClient.getSites("All")
     assert result["OK"] is True, result["Message"]
 
-    assert "testActive1.test.test" in result["Value"]
-    assert "testActive.test.test" in result["Value"]
-    assert "testBanned.test.test" in result["Value"]
-
     # TEST getUsableSites
     # ...............................................................................
 
     result = stClient.getUsableSites()
     assert result["OK"] is True, result["Message"]
 
-    assert "testActive1.test.test" in result["Value"]
-    assert "testActive.test.test" in result["Value"]
-
     # setting a status
-    result = stClient.setSiteStatus("testBanned.test.test", "Probing")
-    assert result["OK"] is True, result["Message"]
-    stClient.rssCache.refreshCache()
+    if inRSS:
+        result = stClient.setSiteStatus("testBanned.test.test", "Probing")
+        assert result["OK"] is True, result["Message"]
+        stClient.rssCache.refreshCache()
 
-    result = stClient.getSites("Probing")
-    assert result["OK"] is True, result["Message"]
-    assert "testBanned.test.test" in result["Value"]
-    assert "testActive.test.test" not in result["Value"]
+        result = stClient.getSites("Probing")
+        assert result["OK"] is True, result["Message"]
